fix: verified user migration in sqlite3
