+++ packages/backend/src/app/database/migration/sqlite/1703713027818-UpdateStatusInUserSqlite.ts
@@ -1,19 +1,19 @@
 import { MigrationInterface, QueryRunner } from 'typeorm'
-import { logger } from '../../helper/logger'
+import { logger } from '../../../helper/logger'
 
-export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInterface {
-    name = 'RenameAppNameToPieceNameSqlite1703713475755'
+export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface {
+    name = 'UpdateStatusInUserSqlite1703713027818'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
-        logger.info('RenameAppNameToPieceNameSqlite1703713475755 up')
+        logger.info('UpdateStatusInUserSqlite1703713027818 up')
         await queryRunner.query(`
             DROP INDEX "idx_user_platform_id_email"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
+            DROP INDEX "idx_user_platform_id_external_id"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_platform_id_external_id"
+            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
         `)
         await queryRunner.query(`
             CREATE TABLE "temporary_user" (
@@ -78,67 +78,17 @@ export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInt
             CREATE UNIQUE INDEX "idx_user_platform_id_email" ON "user" ("platformId", "email")
         `)
         await queryRunner.query(`
+            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
+        `)
+        await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
             WHERE "platformId" IS NULL
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
-        `)
-        await queryRunner.query(`
-            DROP INDEX "idx_app_connection_project_id_and_name"
-        `)
-        await queryRunner.query(`
-            CREATE TABLE "temporary_app_connection" (
-                "id" varchar(21) PRIMARY KEY NOT NULL,
-                "created" datetime NOT NULL DEFAULT (datetime('now')),
-                "updated" datetime NOT NULL DEFAULT (datetime('now')),
-                "name" varchar NOT NULL,
-                "pieceName" varchar NOT NULL,
-                "projectId" varchar(21) NOT NULL,
-                "value" text NOT NULL,
-                "type" varchar NOT NULL,
-                "status" varchar NOT NULL DEFAULT ('ACTIVE'),
-                CONSTRAINT "fk_app_connection_app_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION
-            )
-        `)
-        await queryRunner.query(`
-            INSERT INTO "temporary_app_connection"(
-                    "id",
-                    "created",
-                    "updated",
-                    "name",
-                    "pieceName",
-                    "projectId",
-                    "value",
-                    "type",
-                    "status"
-                )
-            SELECT "id",
-                "created",
-                "updated",
-                "name",
-                "appName",
-                "projectId",
-                "value",
-                "type",
-                "status"
-            FROM "app_connection"
-        `)
-        await queryRunner.query(`
-            DROP TABLE "app_connection"
-        `)
-        await queryRunner.query(`
-            ALTER TABLE "temporary_app_connection"
-                RENAME TO "app_connection"
-        `)
-        await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_app_connection_project_id_and_name" ON "app_connection" ("projectId", "name")
+            DROP INDEX "idx_flow_project_id"
         `)
         await queryRunner.query(`
             DROP INDEX "idx_flow_folder_id"
-        `)
-        await queryRunner.query(`
-            DROP INDEX "idx_flow_project_id"
         `)
         await queryRunner.query(`
             CREATE TABLE "temporary_flow" (
@@ -151,10 +101,10 @@ export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInt
                 "schedule" text,
                 "publishedVersionId" varchar(21),
                 CONSTRAINT "UQ_15375936ad7b8c5dc3f50783a22" UNIQUE ("publishedVersionId"),
-                CONSTRAINT "fk_flow_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION,
+                CONSTRAINT "fk_flow_published_version" FOREIGN KEY ("publishedVersionId") REFERENCES "flow_version" ("id") ON DELETE RESTRICT ON UPDATE NO ACTION,
                 CONSTRAINT "fk_flow_folder_id" FOREIGN KEY ("folderId") REFERENCES "folder" ("id") ON DELETE
                 SET NULL ON UPDATE NO ACTION,
-                    CONSTRAINT "fk_flow_published_version" FOREIGN KEY ("publishedVersionId") REFERENCES "flow_version" ("id") ON DELETE RESTRICT ON UPDATE NO ACTION
+                    CONSTRAINT "fk_flow_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION
             )
         `)
         await queryRunner.query(`
@@ -186,19 +136,19 @@ export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInt
                 RENAME TO "flow"
         `)
         await queryRunner.query(`
-            CREATE INDEX "idx_flow_folder_id" ON "flow" ("folderId")
+            CREATE INDEX "idx_flow_project_id" ON "flow" ("projectId")
         `)
         await queryRunner.query(`
-            CREATE INDEX "idx_flow_project_id" ON "flow" ("projectId")
+            CREATE INDEX "idx_flow_folder_id" ON "flow" ("folderId")
         `)
         await queryRunner.query(`
             DROP INDEX "idx_user_platform_id_email"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
+            DROP INDEX "idx_user_platform_id_external_id"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_platform_id_external_id"
+            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
         `)
         await queryRunner.query(`
             CREATE TABLE "temporary_user" (
@@ -260,11 +210,11 @@ export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInt
                 RENAME TO "user"
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
-            WHERE "platformId" IS NULL
+            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
+            CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
+            WHERE "platformId" IS NULL
         `)
         await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_platform_id_email" ON "user" ("platformId", "email")
@@ -276,10 +226,10 @@ export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInt
             DROP INDEX "idx_user_platform_id_email"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_platform_id_external_id"
+            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
+            DROP INDEX "idx_user_platform_id_external_id"
         `)
         await queryRunner.query(`
             ALTER TABLE "user"
@@ -341,20 +291,20 @@ export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInt
             DROP TABLE "temporary_user"
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
-        `)
-        await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
             WHERE "platformId" IS NULL
+        `)
+        await queryRunner.query(`
+            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
         `)
         await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_platform_id_email" ON "user" ("platformId", "email")
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_flow_project_id"
+            DROP INDEX "idx_flow_folder_id"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_flow_folder_id"
+            DROP INDEX "idx_flow_project_id"
         `)
         await queryRunner.query(`
             ALTER TABLE "flow"
@@ -371,10 +321,10 @@ export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInt
                 "schedule" text,
                 "publishedVersionId" varchar(21),
                 CONSTRAINT "UQ_15375936ad7b8c5dc3f50783a22" UNIQUE ("publishedVersionId"),
-                CONSTRAINT "fk_flow_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION,
+                CONSTRAINT "fk_flow_published_version" FOREIGN KEY ("publishedVersionId") REFERENCES "flow_version" ("id") ON DELETE RESTRICT ON UPDATE NO ACTION,
                 CONSTRAINT "fk_flow_folder_id" FOREIGN KEY ("folderId") REFERENCES "folder" ("id") ON DELETE
                 SET NULL ON UPDATE NO ACTION,
-                    CONSTRAINT "fk_flow_published_version" FOREIGN KEY ("publishedVersionId") REFERENCES "flow_version" ("id") ON DELETE RESTRICT ON UPDATE NO ACTION
+                    CONSTRAINT "fk_flow_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION
             )
         `)
         await queryRunner.query(`
@@ -402,66 +352,16 @@ export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInt
             DROP TABLE "temporary_flow"
         `)
         await queryRunner.query(`
-            CREATE INDEX "idx_flow_project_id" ON "flow" ("projectId")
-        `)
-        await queryRunner.query(`
             CREATE INDEX "idx_flow_folder_id" ON "flow" ("folderId")
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_app_connection_project_id_and_name"
-        `)
-        await queryRunner.query(`
-            ALTER TABLE "app_connection"
-                RENAME TO "temporary_app_connection"
-        `)
-        await queryRunner.query(`
-            CREATE TABLE "app_connection" (
-                "id" varchar(21) PRIMARY KEY NOT NULL,
-                "created" datetime NOT NULL DEFAULT (datetime('now')),
-                "updated" datetime NOT NULL DEFAULT (datetime('now')),
-                "name" varchar NOT NULL,
-                "appName" varchar NOT NULL,
-                "projectId" varchar(21) NOT NULL,
-                "value" text NOT NULL,
-                "type" varchar NOT NULL,
-                "status" varchar NOT NULL DEFAULT ('ACTIVE'),
-                CONSTRAINT "fk_app_connection_app_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION
-            )
-        `)
-        await queryRunner.query(`
-            INSERT INTO "app_connection"(
-                    "id",
-                    "created",
-                    "updated",
-                    "name",
-                    "appName",
-                    "projectId",
-                    "value",
-                    "type",
-                    "status"
-                )
-            SELECT "id",
-                "created",
-                "updated",
-                "name",
-                "pieceName",
-                "projectId",
-                "value",
-                "type",
-                "status"
-            FROM "temporary_app_connection"
-        `)
-        await queryRunner.query(`
-            DROP TABLE "temporary_app_connection"
-        `)
-        await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_app_connection_project_id_and_name" ON "app_connection" ("projectId", "name")
-        `)
-        await queryRunner.query(`
-            DROP INDEX "idx_user_platform_id_external_id"
+            CREATE INDEX "idx_flow_project_id" ON "flow" ("projectId")
         `)
         await queryRunner.query(`
             DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
+        `)
+        await queryRunner.query(`
+            DROP INDEX "idx_user_platform_id_external_id"
         `)
         await queryRunner.query(`
             DROP INDEX "idx_user_platform_id_email"
@@ -526,11 +426,11 @@ export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInt
             DROP TABLE "temporary_user"
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
-        `)
-        await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
             WHERE "platformId" IS NULL
+        `)
+        await queryRunner.query(`
+            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
         `)
         await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_platform_id_email" ON "user" ("platformId", "email")
