+++ packages/backend/src/app/database/migration/sqlite/1703713475755-RenameAppNameToPieceNameSqlite.ts
@@ -1,19 +1,19 @@
 import { MigrationInterface, QueryRunner } from 'typeorm'
-import { logger } from '../../helper/logger'
+import { logger } from '../../../helper/logger'
 
-export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface {
-    name = 'UpdateStatusInUserSqlite1703713027818'
+export class RenameAppNameToPieceNameSqlite1703713475755 implements MigrationInterface {
+    name = 'RenameAppNameToPieceNameSqlite1703713475755'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
-        logger.info('UpdateStatusInUserSqlite1703713027818 up')
+        logger.info('RenameAppNameToPieceNameSqlite1703713475755 up')
         await queryRunner.query(`
             DROP INDEX "idx_user_platform_id_email"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_platform_id_external_id"
+            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
+            DROP INDEX "idx_user_platform_id_external_id"
         `)
         await queryRunner.query(`
             CREATE TABLE "temporary_user" (
@@ -78,17 +78,67 @@ export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface
             CREATE UNIQUE INDEX "idx_user_platform_id_email" ON "user" ("platformId", "email")
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
-        `)
-        await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
             WHERE "platformId" IS NULL
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_flow_project_id"
+            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
+        `)
+        await queryRunner.query(`
+            DROP INDEX "idx_app_connection_project_id_and_name"
+        `)
+        await queryRunner.query(`
+            CREATE TABLE "temporary_app_connection" (
+                "id" varchar(21) PRIMARY KEY NOT NULL,
+                "created" datetime NOT NULL DEFAULT (datetime('now')),
+                "updated" datetime NOT NULL DEFAULT (datetime('now')),
+                "name" varchar NOT NULL,
+                "pieceName" varchar NOT NULL,
+                "projectId" varchar(21) NOT NULL,
+                "value" text NOT NULL,
+                "type" varchar NOT NULL,
+                "status" varchar NOT NULL DEFAULT ('ACTIVE'),
+                CONSTRAINT "fk_app_connection_app_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION
+            )
+        `)
+        await queryRunner.query(`
+            INSERT INTO "temporary_app_connection"(
+                    "id",
+                    "created",
+                    "updated",
+                    "name",
+                    "pieceName",
+                    "projectId",
+                    "value",
+                    "type",
+                    "status"
+                )
+            SELECT "id",
+                "created",
+                "updated",
+                "name",
+                "appName",
+                "projectId",
+                "value",
+                "type",
+                "status"
+            FROM "app_connection"
+        `)
+        await queryRunner.query(`
+            DROP TABLE "app_connection"
+        `)
+        await queryRunner.query(`
+            ALTER TABLE "temporary_app_connection"
+                RENAME TO "app_connection"
+        `)
+        await queryRunner.query(`
+            CREATE UNIQUE INDEX "idx_app_connection_project_id_and_name" ON "app_connection" ("projectId", "name")
         `)
         await queryRunner.query(`
             DROP INDEX "idx_flow_folder_id"
+        `)
+        await queryRunner.query(`
+            DROP INDEX "idx_flow_project_id"
         `)
         await queryRunner.query(`
             CREATE TABLE "temporary_flow" (
@@ -101,10 +151,10 @@ export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface
                 "schedule" text,
                 "publishedVersionId" varchar(21),
                 CONSTRAINT "UQ_15375936ad7b8c5dc3f50783a22" UNIQUE ("publishedVersionId"),
-                CONSTRAINT "fk_flow_published_version" FOREIGN KEY ("publishedVersionId") REFERENCES "flow_version" ("id") ON DELETE RESTRICT ON UPDATE NO ACTION,
+                CONSTRAINT "fk_flow_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION,
                 CONSTRAINT "fk_flow_folder_id" FOREIGN KEY ("folderId") REFERENCES "folder" ("id") ON DELETE
                 SET NULL ON UPDATE NO ACTION,
-                    CONSTRAINT "fk_flow_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION
+                    CONSTRAINT "fk_flow_published_version" FOREIGN KEY ("publishedVersionId") REFERENCES "flow_version" ("id") ON DELETE RESTRICT ON UPDATE NO ACTION
             )
         `)
         await queryRunner.query(`
@@ -136,19 +186,19 @@ export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface
                 RENAME TO "flow"
         `)
         await queryRunner.query(`
-            CREATE INDEX "idx_flow_project_id" ON "flow" ("projectId")
+            CREATE INDEX "idx_flow_folder_id" ON "flow" ("folderId")
         `)
         await queryRunner.query(`
-            CREATE INDEX "idx_flow_folder_id" ON "flow" ("folderId")
+            CREATE INDEX "idx_flow_project_id" ON "flow" ("projectId")
         `)
         await queryRunner.query(`
             DROP INDEX "idx_user_platform_id_email"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_platform_id_external_id"
+            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
+            DROP INDEX "idx_user_platform_id_external_id"
         `)
         await queryRunner.query(`
             CREATE TABLE "temporary_user" (
@@ -210,11 +260,11 @@ export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface
                 RENAME TO "user"
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
-        `)
-        await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
             WHERE "platformId" IS NULL
+        `)
+        await queryRunner.query(`
+            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
         `)
         await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_platform_id_email" ON "user" ("platformId", "email")
@@ -226,10 +276,10 @@ export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface
             DROP INDEX "idx_user_platform_id_email"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
+            DROP INDEX "idx_user_platform_id_external_id"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_platform_id_external_id"
+            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
         `)
         await queryRunner.query(`
             ALTER TABLE "user"
@@ -291,20 +341,20 @@ export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface
             DROP TABLE "temporary_user"
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
-            WHERE "platformId" IS NULL
+            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
+            CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
+            WHERE "platformId" IS NULL
         `)
         await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_platform_id_email" ON "user" ("platformId", "email")
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_flow_folder_id"
+            DROP INDEX "idx_flow_project_id"
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_flow_project_id"
+            DROP INDEX "idx_flow_folder_id"
         `)
         await queryRunner.query(`
             ALTER TABLE "flow"
@@ -321,10 +371,10 @@ export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface
                 "schedule" text,
                 "publishedVersionId" varchar(21),
                 CONSTRAINT "UQ_15375936ad7b8c5dc3f50783a22" UNIQUE ("publishedVersionId"),
-                CONSTRAINT "fk_flow_published_version" FOREIGN KEY ("publishedVersionId") REFERENCES "flow_version" ("id") ON DELETE RESTRICT ON UPDATE NO ACTION,
+                CONSTRAINT "fk_flow_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION,
                 CONSTRAINT "fk_flow_folder_id" FOREIGN KEY ("folderId") REFERENCES "folder" ("id") ON DELETE
                 SET NULL ON UPDATE NO ACTION,
-                    CONSTRAINT "fk_flow_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION
+                    CONSTRAINT "fk_flow_published_version" FOREIGN KEY ("publishedVersionId") REFERENCES "flow_version" ("id") ON DELETE RESTRICT ON UPDATE NO ACTION
             )
         `)
         await queryRunner.query(`
@@ -352,16 +402,66 @@ export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface
             DROP TABLE "temporary_flow"
         `)
         await queryRunner.query(`
-            CREATE INDEX "idx_flow_folder_id" ON "flow" ("folderId")
-        `)
-        await queryRunner.query(`
             CREATE INDEX "idx_flow_project_id" ON "flow" ("projectId")
         `)
         await queryRunner.query(`
-            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
+            CREATE INDEX "idx_flow_folder_id" ON "flow" ("folderId")
+        `)
+        await queryRunner.query(`
+            DROP INDEX "idx_app_connection_project_id_and_name"
+        `)
+        await queryRunner.query(`
+            ALTER TABLE "app_connection"
+                RENAME TO "temporary_app_connection"
+        `)
+        await queryRunner.query(`
+            CREATE TABLE "app_connection" (
+                "id" varchar(21) PRIMARY KEY NOT NULL,
+                "created" datetime NOT NULL DEFAULT (datetime('now')),
+                "updated" datetime NOT NULL DEFAULT (datetime('now')),
+                "name" varchar NOT NULL,
+                "appName" varchar NOT NULL,
+                "projectId" varchar(21) NOT NULL,
+                "value" text NOT NULL,
+                "type" varchar NOT NULL,
+                "status" varchar NOT NULL DEFAULT ('ACTIVE'),
+                CONSTRAINT "fk_app_connection_app_project_id" FOREIGN KEY ("projectId") REFERENCES "project" ("id") ON DELETE CASCADE ON UPDATE NO ACTION
+            )
+        `)
+        await queryRunner.query(`
+            INSERT INTO "app_connection"(
+                    "id",
+                    "created",
+                    "updated",
+                    "name",
+                    "appName",
+                    "projectId",
+                    "value",
+                    "type",
+                    "status"
+                )
+            SELECT "id",
+                "created",
+                "updated",
+                "name",
+                "pieceName",
+                "projectId",
+                "value",
+                "type",
+                "status"
+            FROM "temporary_app_connection"
+        `)
+        await queryRunner.query(`
+            DROP TABLE "temporary_app_connection"
+        `)
+        await queryRunner.query(`
+            CREATE UNIQUE INDEX "idx_app_connection_project_id_and_name" ON "app_connection" ("projectId", "name")
         `)
         await queryRunner.query(`
             DROP INDEX "idx_user_platform_id_external_id"
+        `)
+        await queryRunner.query(`
+            DROP INDEX "idx_user_partial_unique_email_platform_id_is_null"
         `)
         await queryRunner.query(`
             DROP INDEX "idx_user_platform_id_email"
@@ -426,11 +526,11 @@ export class UpdateStatusInUserSqlite1703713027818 implements MigrationInterface
             DROP TABLE "temporary_user"
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
-            WHERE "platformId" IS NULL
+            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
         `)
         await queryRunner.query(`
-            CREATE UNIQUE INDEX "idx_user_platform_id_external_id" ON "user" ("platformId", "externalId")
+            CREATE UNIQUE INDEX "idx_user_partial_unique_email_platform_id_is_null" ON "user" ("email")
+            WHERE "platformId" IS NULL
         `)
         await queryRunner.query(`
             CREATE UNIQUE INDEX "idx_user_platform_id_email" ON "user" ("platformId", "email")
