+++ packages/backend/src/app/database/sqlite-connection.ts
@@ -27,9 +27,9 @@ import { AddPlatformIdToPieceMetadataSqlite1700524446967 } from './migration/sql
 import { AddPartialUniqueIndexForEmailAndPlatformIdIsNull1701096458822 } from './migration/common/1701096458822-add-partial-unique-index-for-email-and-platform-id-is-null'
 import { AddPlatformIdToFileSqlite1701808264444 } from './migration/sqlite/1701808264444-AddPlatformIdToFileSqlite'
 import { RemoveFlowInstanceSqlite1702412280963 } from './migration/sqlite/1702412280963-remove-flow-instance-sqlite'
-import { UpdateStatusInUserSqlite1703713027818 } from './migration/1703713027818-UpdateStatusInUserSqlite'
-import { RenameAppNameToPieceNameSqlite1703713475755 } from './migration/1703713475755-RenameAppNameToPieceNameSqlite'
-import { AddVerifiedAndChangeStatusSqlite1703768553820 } from './migration/1703768553820-AddVerifiedAndChangeStatusSqlite'
+import { UpdateStatusInUserSqlite1703713027818 } from './migration/sqlite/1703713027818-UpdateStatusInUserSqlite'
+import { RenameAppNameToPieceNameSqlite1703713475755 } from './migration/sqlite/1703713475755-RenameAppNameToPieceNameSqlite'
+import { AddVerifiedAndChangeStatusSqlite1703768553820 } from './migration/sqlite/1703768553820-AddVerifiedAndChangeStatusSqlite'
 
 const getSqliteDatabaseFilePath = (): string => {
     const apConfigDirectoryPath = system.getOrThrow(SystemProp.CONFIG_PATH)
